requirements.txt
telethon==1.36.0
discord.py==2.3.2
audioop-lts; python_version >= "3.13"
python-dotenv==1.0.1
aiohttp==3.11.11
aiofiles==23.2.1
Pillow>=11.1.0